Merge/integrate UTF-8 support for XFree86 and Xft using code from fltk_utf8. (still need to add X11, OSX, and WIN32 support, and re-merge input/text changes for composition "spot" stuff...)
Drop XUtf8 library, since we don't need it specifically; we can use a much
smaller subset of the current code...


git-svn-id: file:///fltk/svn/branches/branch-1.2@3179 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/Fl_Tile.cxx
@@ -1,9 +1,9 @@
 //
-// "$Id: Fl_Tile.cxx,v 1.5.2.5.2.3.2.2 2003/11/07 03:47:24 easysw Exp $"
+// "$Id: Fl_Tile.cxx,v 1.5.2.5.2.3.2.3 2003/12/02 02:51:47 easysw Exp $"
 //
 // Tile widget for the Fast Light Tool Kit (FLTK).
 //
-// Copyright 1998-2004 by Bill Spitzak and others.
+// Copyright 1998-2003 by Bill Spitzak and others.
 //
 // This library is free software; you can redistribute it and/or
 // modify it under the terms of the GNU Library General Public
@@ -37,7 +37,7 @@
 
 void Fl_Tile::position(int oix, int oiy, int newx, int newy) {
   Fl_Widget*const* a = array();
-  int* p = sizes();
+  short* p = sizes();
   p += 8; // skip group & resizable's saved size
   for (int i=children(); i--; p += 4) {
     Fl_Widget* o = *a++;
@@ -69,7 +69,7 @@ void Fl_Tile::resize(int X,int Y,int W,int H) {
   int dy = Y-y();
   int dw = W-w();
   int dh = H-h();
-  int* p = sizes();
+  short* p = sizes();
   // resize this (skip the Fl_Group resize):
   Fl_Widget::resize(X,Y,W,H);
   // find bottom-right of resiable:
@@ -132,8 +132,8 @@ int Fl_Tile::handle(int event) {
     int oldx = 0;
     int oldy = 0;
     Fl_Widget*const* a = array();
-    int* q = sizes();
-    int* p = q+8;
+    short* q = sizes();
+    short* p = q+8;
     for (int i=children(); i--; p += 4) {
       Fl_Widget* o = *a++;
       if (o == resizable()) continue;
@@ -196,5 +196,5 @@ int Fl_Tile::handle(int event) {
 }
 
 //
-// End of "$Id: Fl_Tile.cxx,v 1.5.2.5.2.3.2.2 2003/11/07 03:47:24 easysw Exp $".
+// End of "$Id: Fl_Tile.cxx,v 1.5.2.5.2.3.2.3 2003/12/02 02:51:47 easysw Exp $".
 //
