+++ test/arc.cxx
@@ -1,9 +1,9 @@
 //
-// "$Id: arc.cxx,v 1.4.2.3.2.1.2.2 2003/11/07 03:47:25 easysw Exp $"
+// "$Id: arc.cxx,v 1.4.2.3.2.1.2.3 2003/12/02 02:51:49 easysw Exp $"
 //
 // Arc drawing test program for the Fast Light Tool Kit (FLTK).
 //
-// Copyright 1998-2004 by Bill Spitzak and others.
+// Copyright 1998-2003 by Bill Spitzak and others.
 //
 // This library is free software; you can redistribute it and/or
 // modify it under the terms of the GNU Library General Public
@@ -82,7 +82,7 @@ int main(int argc, char** argv) {
     s->step(1);
     s->value(args[n]);
     s->align(FL_ALIGN_LEFT);
-    s->callback(slider_cb, (void*)((long)n));
+    s->callback(slider_cb, (void*)n);
   }
 
   window.end();
@@ -92,6 +92,6 @@ int main(int argc, char** argv) {
 
 
 //
-// End of "$Id: arc.cxx,v 1.4.2.3.2.1.2.2 2003/11/07 03:47:25 easysw Exp $".
+// End of "$Id: arc.cxx,v 1.4.2.3.2.1.2.3 2003/12/02 02:51:49 easysw Exp $".
 //
